# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrack API
  description: >
    Production REST API for FinTrack Personal Financial Management and Asset Tracking.
    Implements cash flow tracking, dynamic budgeting, income allocation, customizable
    financial cycles, receivables (loans), investments, and periodic reporting.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker Compose Environment
paths:
...
  /auth/register:
    post:
      summary: User Registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, example: "user@example.com" }
                password: { type: string, minLength: 8, example: "SecurePass123!" }
                cycle_start_day: { type: integer, minimum: 1, maximum: 31, default: 1 }
      responses:
        '201':
          description: User created
        '400':
          description: Validation error
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: User Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Authentication successful, returns JWT token
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      summary: Current User Profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile details
  /auth/settings:
    put:
      summary: Update User Cycle Settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cycle_start_day]
              properties:
                cycle_start_day: { type: integer, minimum: 1, maximum: 31 }
      responses:
        '200':
          description: Settings updated
  /accounts:
    get:
      summary: List Accounts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User accounts list
    post:
      summary: Create Account
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string }
                type: { type: string, enum: [BANK, EWALLET, CASH, OTHER] }
                opening_balance: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Account created
...
  /allocations:
    get:
      summary: List Allocations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Income allocations for cycle
    post:
      summary: Create Income Allocation
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Allocation created
        '422':
          description: Allocations exceed total cycle income
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT